added try catch and error messages

+++ DockerGUI/Form1.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Diagnostics;
 using System.IO;
-using System.Management.Automation;
-using System.Management.Automation.Runspaces;
+using System.Linq;
 using System.Windows.Forms;
-using PowerShell = System.Management.Automation.PowerShell;
 
 namespace DockerGUI
 {
@@ -38,6 +36,9 @@ namespace DockerGUI
             String root = @"C:\DockerGUI";
             String subdir = @"C:\DockerGUI\Dockerfiles";
             String fileName = "";
+            bool isPhp = false;
+            bool isJava = false;
+            bool isCsharp = false;
             // If the root directory doesn't exsist the root directory is created
             if (!Directory.Exists(root))
             {
@@ -58,20 +59,70 @@ namespace DockerGUI
                 {
                     Console.WriteLine("File has been recived");
                     fileName = Path.GetFileName(fbd.SelectedPath);
+                    try
+                    {
+                        // Reads the contents of the files in the directory that are php files and looks for any files with ".blade" in its contents. This will be used to identify if a project is using the Laravel
+                        // framework so the propper dependancies can be added. 
+                        bool hasLaravel = false;
+                        var files = from file in Directory.EnumerateFiles(fbd.SelectedPath, "*.php", SearchOption.AllDirectories)
+                                    from line in File.ReadLines(file)
+                                    where file.Contains(".php")
+                                    select new
+                                    {
+                                        File = file,
+                                        Line = line
+                                    };
+                        foreach (var f in files)
+                        {
+                            isPhp = true;
+                        }
+                        files = from file in Directory.EnumerateFiles(fbd.SelectedPath, "*.php", SearchOption.AllDirectories)
+                                from line in File.ReadLines(file)
+                                where line.Contains(".blade")
+                                select new
+                                {
+                                    File = file,
+                                    Line = line
+                                };
+                        foreach (var f in files)
+                        {
+                            hasLaravel = true;
+                        }
+                        if (isPhp == true)
+                        {
+                            // Create an empty process 
+                            Process process;
+                            // Set the process to be a single instance of Powershell with a Docker command
+                            ProcessStartInfo processInfo = new ProcessStartInfo($@"C:\Windows\system32\WindowsPowerShell\v1.0\powershell.exe", $@"docker run -it --rm -p 8081:8080 -e PASSWORD=test -v " + fbd.SelectedPath + " --name vscode codercom/code-server");
+
+                            // Parameters for the process instance, CreateNoWindow is commented out for testing purposes
+                            //processInfo.CreateNoWindow = true;
+                            processInfo.UseShellExecute = false;
+                            process = Process.Start(processInfo);
+
+                            //the line below locks the original form window until the powershell window is closed
+                            process.WaitForExit();
+                            process.Close();
+                        }
+                        else
+                        {
+                            string UFTmessage = "This file is unsupported by this application, please select a new file.";
+                            string UFTcaption = "Unsupported File Type";
+                            MessageBoxButtons UFTbutton = MessageBoxButtons.OK;
+                            MessageBox.Show(UFTmessage, UFTcaption, UFTbutton);
+                            MessageBox.Show(UFTmessage, UFTcaption, UFTbutton);
+                        }
+                    }
+                    
+                    catch (UnauthorizedAccessException uAEx)
+                        {
+                    Console.WriteLine(uAEx.Message);
+                        }
+                    catch (PathTooLongException pathEx)
+                        {
+                    Console.WriteLine(pathEx.Message);
+                        }
                 }
-                // Create an empty process 
-                Process process;
-                // Set the process to be a single instance of Powershell with a Docker command
-                ProcessStartInfo processInfo = new ProcessStartInfo($@"C:\Windows\system32\WindowsPowerShell\v1.0\powershell.exe", $@"docker run -it --rm -p 8081:8080 -e PASSWORD=test -v " + fbd.SelectedPath + " --name vscode codercom/code-server");
-
-                // Parameters for the process instance, CreateNoWindow is commented out for testing purposes
-                //processInfo.CreateNoWindow = true;
-                processInfo.UseShellExecute = false;
-                process = Process.Start(processInfo);
-
-                //the line below locks the original form window until the powershell window is closed
-                process.WaitForExit();
-                process.Close();
             }
         }
         private void MainPage_Load(object sender, EventArgs e)
